changed rates to float32 to support 1.6x exp and stuff

+++ channel/server.go
@@ -74,9 +74,9 @@ func (p *players) removeFromConn(conn mnet.Client) error {
 }
 
 type rates struct {
-	exp   int16
-	drop  int16
-	mesos int16
+	exp   float32
+	drop  float32
+	mesos float32
 }
 
 // Server state
