+++ server_config.go
@@ -28,9 +28,9 @@ type loginConfig struct {
 type worldConfig struct {
 	Message         string
 	Ribbon          byte
-	ExpRate         int
-	DropRate        int
-	MesosRate       int
+	ExpRate         float32
+	DropRate        float32
+	MesosRate       float32
 	LoginAddress    string
 	LoginPort       string
 	ListenAddress   string
